+++ nucypher-core-python/src/lib.rs
@@ -643,7 +643,7 @@ pub struct RequestSharedSecret {
 }
 
 #[pyclass(module = "nucypher_core")]
-#[derive(derive_more::From, derive_more::AsRef)]
+#[derive(Clone, PartialEq, Eq, derive_more::From, derive_more::AsRef)]
 pub struct RequestPublicKey {
     backend: nucypher_core::RequestPublicKey,
 }
@@ -657,6 +657,14 @@ impl RequestPublicKey {
 
     fn __bytes__(&self) -> PyObject {
         to_bytes(self)
+    }
+
+    fn __richcmp__(&self, other: &Self, op: CompareOp) -> PyResult<bool> {
+        richcmp(self, other, op)
+    }
+
+    fn __hash__(&self) -> PyResult<isize> {
+        hash("RequestPublicKey", self)
     }
 
     fn __str__(&self) -> PyResult<String> {
@@ -679,7 +687,6 @@ impl RequestSecretKey {
         })
     }
 
-    #[getter]
     pub fn public_key(&self) -> RequestPublicKey {
         RequestPublicKey {
             backend: self.backend.public_key(),
@@ -831,10 +838,11 @@ impl ThresholdDecryptionRequest {
         shared_secret: &RequestSharedSecret,
         requester_public_key: &RequestPublicKey,
     ) -> EncryptedThresholdDecryptionRequest {
+        let encrypted_request = self
+            .backend
+            .encrypt(shared_secret.as_ref(), requester_public_key.as_ref());
         EncryptedThresholdDecryptionRequest {
-            backend: self
-                .backend
-                .encrypt(shared_secret.as_ref(), requester_public_key.as_ref()),
+            backend: encrypted_request,
         }
     }
 
@@ -865,8 +873,9 @@ impl EncryptedThresholdDecryptionRequest {
         self.backend.ritual_id
     }
 
+    #[getter]
     pub fn requester_public_key(&self) -> RequestPublicKey {
-        RequestPublicKey::from(self.backend.requester_public_key)
+        self.backend.requester_public_key.into()
     }
 
     pub fn decrypt(
@@ -902,10 +911,15 @@ pub struct ThresholdDecryptionResponse {
 #[pymethods]
 impl ThresholdDecryptionResponse {
     #[new]
-    pub fn new(decryption_share: &[u8]) -> Self {
+    pub fn new(ritual_id: u16, decryption_share: &[u8]) -> Self {
         ThresholdDecryptionResponse {
-            backend: nucypher_core::ThresholdDecryptionResponse::new(decryption_share),
+            backend: nucypher_core::ThresholdDecryptionResponse::new(ritual_id, decryption_share),
         }
+    }
+
+    #[getter]
+    pub fn ritual_id(&self) -> u16 {
+        self.backend.ritual_id
     }
 
     #[getter]
@@ -944,6 +958,11 @@ pub struct EncryptedThresholdDecryptionResponse {
 
 #[pymethods]
 impl EncryptedThresholdDecryptionResponse {
+    #[getter]
+    pub fn ritual_id(&self) -> u16 {
+        self.backend.ritual_id
+    }
+
     pub fn decrypt(
         &self,
         shared_secret: &RequestSharedSecret,